Tests sur un Debian (Raspberry) & corrections de bugs

+++ include/position.h
@@ -19,31 +19,31 @@ Param setLorenz(Param_Lorenz L, Param P);
 Param setVanDerPol(Param_VanDerPol V, Param P);
 Param setRossler(Param_Rossler R, Param P);
 
-Param setParamLorenz(Param param, float B, float S, float P);
-Param setParamVanDerPol(Param param, float K, float M, float B, float S, float P, float Q);
-Param setParamRossler(Param param, float A, float B, float C);
+Param setParamLorenz(Param param, double B, double S, double P);
+Param setParamVanDerPol(Param param, double K, double M, double B, double S, double P, double Q);
+Param setParamRossler(Param param, double A, double B, double C);
 
 //Recupere la valeur de x de la position
-float get_x (Coord position);
+double get_x (Coord position);
 
 //Recupere la valeur de y de la position
-float get_y (Coord position);
+double get_y (Coord position);
 
 //Recupere la valeur de z de la position
-float get_z(Coord position);
+double get_z(Coord position);
 
 //Recupere la valeur de t de la position
-float get_t (Coord position);
+double get_t (Coord position);
 
 
 //Création d'une nouvelle coordonnees
-Coord new_coord(float t, float x, float y, float z);
+Coord new_coord(double t, double x, double y, double z);
 
 //Calcul de la nouvelle position - Attracteur de Lorenz
-Coord position_next_Lorenz(Coord point, float dt, float B, float P, float S);
+Coord position_next_Lorenz(Coord point, double dt, double B, double P, double S);
 
 //Calcul de la nouvelle position - Attracteur de Van der Pol
-Coord position_next_VanDerPol(Coord point, float dt, float K, float M, float B, float S, float P, float Q);
+Coord position_next_VanDerPol(Coord point, double dt, double K, double M, double B, double S, double P, double Q);
 
 //Calcul de la nouvelle position - Attracteur de Rössler
-Coord position_next_Rossler(Coord point, float dt, float A, float B,float C);
+Coord position_next_Rossler(Coord point, double dt, double A, double B,double C);
